// DXC_Frontend/src/api/models/feedbackDetailDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */
import type { FeedbackAttachmentDto } from './feedbackAttachmentDto'
import type { FeedbackProcessingDto } from './feedbackProcessingDto'
import type { FeedbackResponseDto } from './feedbackResponseDto'

export interface FeedbackDetailDto {
  id?: number
  publicId?: string
  /**
   * @minLength 1
   * @maxLength 500
   */
  title: string
  /** @minLength 1 */
  content: string
  /**
   * @minLength 1
   * @maxLength 200
   */
  fullName: string
  /**
   * @maxLength 20
   * @nullable
   */
  phoneNumber?: string | null
  /**
   * @maxLength 500
   * @nullable
   */
  location?: string | null
  isPublic?: boolean
  currentStatusId?: number
  /** @nullable */
  currentStatusCode: string | null
  /** @nullable */
  currentStatusName: string | null
  /** @nullable */
  currentStatusColor?: string | null
  /** @nullable */
  assignedDepartmentId?: number | null
  /** @nullable */
  assignedDepartmentPublicId?: string | null
  /** @nullable */
  assignedDepartmentCode?: string | null
  /** @nullable */
  assignedDepartmentName?: string | null
  /** @nullable */
  assignedDepartmentContactEmail?: string | null
  /** @nullable */
  assignedDepartmentContactPhone?: string | null
  isActive?: boolean
  createdAt?: string
  updatedAt?: string
  /** @nullable */
  attachments?: FeedbackAttachmentDto[] | null
  /** @nullable */
  processings?: FeedbackProcessingDto[] | null
  /** @nullable */
  responses?: FeedbackResponseDto[] | null
}